Toolbar, Infobar, EditorEvents, ECS bug fixes, Editor Modules Ordering

+++ ImpulseEditor/ImpulseEditor/Source/Shared/ImpulseEditor.cpp
@@ -22,6 +22,8 @@ ExampleLayer::ExampleLayer()
 ExampleLayer::~ExampleLayer()
 {
 
+
+
 }
 
 void ExampleLayer::OnImGuiRender()
@@ -210,9 +212,9 @@ void ImpulseEditor::OnImGuiSetup()
 	fd = GEngine::FileSystem::FileDataFromPath("EngineContent/Fonts/roboto.ttf");
 	ImFontConfig cfg = {  };
 	cfg.FontDataOwnedByAtlas = false;
-	mainFont = io.Fonts->AddFontFromMemoryTTF(fd->GetData(), fd->GetDataSize(), 20, &cfg);
-	smallFont = io.Fonts->AddFontFromMemoryTTF(fd->GetData(), fd->GetDataSize(), 17, &cfg);
-	largeFont = io.Fonts->AddFontFromMemoryTTF(fd->GetData(), fd->GetDataSize(), 23, &cfg);
+	mainFont = io.Fonts->AddFontFromMemoryTTF(fd->GetData(), (int)fd->GetDataSize(), 20, &cfg);
+	smallFont = io.Fonts->AddFontFromMemoryTTF(fd->GetData(), (int)fd->GetDataSize(), 17, &cfg);
+	largeFont = io.Fonts->AddFontFromMemoryTTF(fd->GetData(), (int)fd->GetDataSize(), 23, &cfg);
 	io.Fonts->Build();
 #endif
 }
